Store old style for animating ::backdrop on anchored elements
For elements that can be transitioned, and have its style updated
multiple times per lifecycle updates, like for container queries, anchor
positioning, etc, we need to store the original old style used for
triggering transitions.

We did so for anchored elements since they can have multiple passes
position fallback evaluation, but not for its pseudo elements. Even if
allowed properties for @position-try blocks do not typically inherit, a
recalc of an anchored element that generates pseudo elements will cause
an update to its pseudo elements for each pass.

That means we need to store the old style for animating pseudo elements
that are generated from anchored originating elements for transitions to
work properly.

Bug: 347967483
Change-Id: Ied0039ce4c95c98f99583601f29873d4562f6527